+++ src/Response/ProxyServiceValidate.php
@@ -10,43 +10,30 @@ declare(strict_types=1);
 namespace EcPhp\CasLib\Response;
 
 use EcPhp\CasLib\Response\Contract\ProxyServiceValidate as ProxyServiceValidateInterface;
-use Psr\Http\Message\StreamInterface;
+use Exception;
+use Psr\Cache\CacheItemPoolInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamFactoryInterface;
+use Psr\Log\LoggerInterface;
 
 final class ProxyServiceValidate extends CasResponse implements ProxyServiceValidateInterface
 {
-    public function getBody()
+    private CacheItemPoolInterface $cache;
+
+    public function __construct(ResponseInterface $response, string $format, CacheItemPoolInterface $cache, StreamFactoryInterface $streamFactory, LoggerInterface $logger)
     {
-        return $this->response->getBody();
+        parent::__construct($response, $format, $streamFactory, $logger);
+        $this->cache = $cache;
+    }
+
+    public function getCache(): CacheItemPoolInterface
+    {
+        return $this->cache;
     }
 
     public function getCredentials(): array
     {
         return $this->parse()['serviceResponse']['authenticationSuccess'];
-    }
-
-    public function getFormat(): string
-    {
-        return $this->format;
-    }
-
-    public function getHeader($name)
-    {
-        return $this->response->getHeader($name);
-    }
-
-    public function getHeaderLine($name)
-    {
-        return $this->response->getHeaderLine($name);
-    }
-
-    public function getHeaders()
-    {
-        return $this->response->getHeaders();
-    }
-
-    public function getProtocolVersion()
-    {
-        return $this->response->getProtocolVersion();
     }
 
     public function getProxies(): array
@@ -58,66 +45,26 @@ final class ProxyServiceValidate extends CasResponse implements ProxyServiceVali
             [];
     }
 
-    public function getReasonPhrase()
+    public function withPgtIou(): ProxyServiceValidateInterface
     {
-        return $this->response->getReasonPhrase();
+        $pgt = $this->parse()['serviceResponse']['authenticationSuccess']['proxyGrantingTicket'];
+
+        if (false === $this->getCache()->hasItem($pgt)) {
+            throw new Exception('CAS validation failed: pgtIou not found in the cache.');
+        }
+
+        return $this
+            ->withBody(
+                $this
+                    ->getStreamFactory()
+                    ->createStream(
+                        str_replace(
+                            $pgt,
+                            $this->getCache()->getItem($pgt)->get(),
+                            (string) $this->getBody()
+                        )
+                    )
+            );
     }
 
-    public function getStatusCode()
-    {
-        return $this->response->getStatusCode();
-    }
-
-    public function hasHeader($name)
-    {
-        return $this->response->hasHeader($name);
-    }
-
-    public function withAddedHeader($name, $value)
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withAddedHeader($name, $value);
-
-        return $clone;
-    }
-
-    public function withBody(StreamInterface $body)
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withBody($body);
-
-        return $clone;
-    }
-
-    public function withHeader($name, $value)
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withHeader($name, $value);
-
-        return $clone;
-    }
-
-    public function withoutHeader($name)
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withoutHeader($name);
-
-        return $clone;
-    }
-
-    public function withProtocolVersion($version)
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withProtocolVersion($version);
-
-        return $clone;
-    }
-
-    public function withStatus($code, $reasonPhrase = '')
-    {
-        $clone = clone $this;
-        $clone->response = $this->response->withStatus($code, $reasonPhrase);
-
-        return $clone;
-    }
 }
